allow fetching postgres bastion keys via data api
This uses a new `/bastion` API on shogun for postgres that exposes
the bastion configuration

This allows us to get rid of bastion key rolling, which is quite
disruptive for customers. There are some tradeoffs:
- This means that `heroku pg:psql` and similar commands, and `heroku
redis:cli` don't work if shogun is down. Shogun's availability is
strictly smaller than the heroku API's (because it relies on the latter
for auth). There is also a risk here that if shogun or API is down (but
the dbs are up) we cannot use this tooling to get into the databases.
However, a thing we can do (and did during the S3 outage last year) is
getting access to a running dogwood container in the respective space,
then using `psql` from there.
- A small number of customers may be using these bastion keys for their
own purposes. We can announce this change through changelog and
whitelist customers (via a formation attr) until we reach some agreed
upon date for deprecation.

The upsides here are notable - with this change there are no more
config var changes for bastion key rolls.
This has (in particular) been hurting Caesar's
(https://trello.com/c/X6npqMGs/1181-allow-setting-a-bastion-key-roll-window).
This PR includes an `attr` we can directly set on Caesar's formations so
they no longer receive these rolls at all.

This change is behind a `FeatureFlag`, and two per formation attrs that
we can use to whitelist customers in or out of this behaviour. This
doesn't change the behaviour for any database that is compliance
related, as doing so could potentially require changing our compliance
docs and agreements with auditors. That seems like a worthy followup,
but it's not worth doing in this first part.